Add author association to campgrounds; implement author checks for editing and deleting; update seed data to include author

+++ YelpCamp/routes/campgrounds.js
@@ -35,6 +35,7 @@ router.post(
   catchAsync(async (req, res) => {
     //if(!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', 'Successfully made a new campground!');
     res.redirect(`campgrounds/${campground._id}`);
@@ -44,9 +45,10 @@ router.post(
 router.get(
   '/:id',
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate(
-      'reviews'
-    );
+    const campground = await Campground.findById(req.params.id)
+      .populate('reviews')
+      .populate('author');
+    console.log(campground);
     if (!campground) {
       req.flash('error', 'Cannot find that campground!');
       return res.redirect('/campgrounds');
@@ -59,10 +61,15 @@ router.get(
   '/:id/edit',
   isLoggedIn,
   catchAsync(async (req, res) => {
+    const { id } = req.params;
     const campground = await Campground.findById(req.params.id);
     if (!campground) {
       req.flash('error', 'Cannot find that campground!');
       return res.redirect('/campgrounds');
+    }
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that');
+      return res.redirect(`/campgrounds/${id}`);
     }
     res.render('campgrounds/edit', { campground });
   })
@@ -74,7 +81,12 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     req.flash('success', 'Successfully updated campground!');
@@ -87,6 +99,11 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted the campground!');
     res.redirect('/campgrounds');
